+++ include/marge.hpp
@@ -19,7 +19,7 @@ class Solver {
 
         void init();
 
-        std::vector<Path> solve(std::string origin, std::string destination, double dt_start, double dt_promise);
+        std::vector<Path> solve(std::string origin, std::string destination, long dt_start, long dt_promise);
 
         std::tuple<std::shared_ptr<Connection>, std::shared_ptr<DeliveryCenter>, std::shared_ptr<DeliveryCenter> > lookup(std::string connection);
 };
